refactor(services): slim mercury/world_data/mod.rs by extracting siblings
Move the test module out of mod.rs into a sibling tests.rs file. mod.rs
now contains just submodule glue, re-exports, and the shared world data
lookup helpers used by both phases.rs and map_loaded.rs.

Public API unchanged. No behavior changes.
